refactor: classes restructured to improve solution structure
