+++ C#/JsonSerialization/Model/Controllers/CarSerializer.cs
@@ -1,11 +1,10 @@
 ﻿using JsonSerialization.Model;
 using System.Text.Json;
-using System;
 
-namespace JsonSerialization {
+namespace JsonSerialization.Model.Controllers {
     public class CarSerializer {
         public Car Deserialize(string jsonObject) {
-            if(jsonObject.Equals("{}")) return null;
+            if (jsonObject.Equals("{}")) return null;
             return JsonSerializer.Deserialize<Car>(jsonObject);
         }
     }
