+++ C#/JsonSerialization/Model/DataCategories/Extra.cs
